This works.
I didn't actually use my total calories function in the avg calories
function, because by the time I thought of that I was already done. I see now
that I could have been clever that way, but for this assignment, it was
just about the same in terms of code and runtime probably so it I
don't think it is worth going back to change it.